+++ src/test/java/com/picimako/mockitools/intention/convert/verification/ConvertBDDMockitoThenToMockitoVerifyIntentionTest.java
@@ -1,0 +1,219 @@
+//Copyright 2021 Tamás Balog. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
+
+package com.picimako.mockitools.intention.convert.verification;
+
+import com.intellij.codeInsight.intention.IntentionAction;
+
+import com.picimako.mockitools.intention.MockitoolsIntentionTestBase;
+
+/**
+ * Functional test for {@link ConvertBDDMockitoThenToMockitoVerifyIntention}.
+ */
+public class ConvertBDDMockitoThenToMockitoVerifyIntentionTest extends MockitoolsIntentionTestBase {
+
+    @Override
+    protected IntentionAction getIntention() {
+        return new ConvertBDDMockitoThenToMockitoVerifyIntention();
+    }
+
+    //Availability
+
+    public void testNotAvailableForNonJavaFile() {
+        checkIntentionIsNotAvailable("NotJava.xml", "<tag><caret></tag>");
+    }
+
+    public void testNotAvailableForNonMethodCallIdentifier() {
+        checkIntentionIsNotAvailable("NotAvailable.java",
+            "class NotAvailable {\n" +
+                "    private String fiel<caret>d;\n" +
+                "}");
+    }
+
+    public void testNotAvailableOnNonBDDMockitoThenMethod() {
+        checkIntentionIsNotAvailable("NotAvailable.java",
+            "import org.mockito.Mockito;\n" +
+                "\n" +
+                "class NotAvailable {\n" +
+                "    void testMethod(){\n" +
+                "        Mockito.mo<caret>ck(Object.class);\n" +
+                "    }\n" +
+                "}");
+    }
+
+    public void testNotAvailableWhenBDDMockitoThenHasNoSubsequentMethodCall() {
+        checkIntentionIsNotAvailable("NotAvailable.java",
+            "import org.mockito.BDDMockito;\n" +
+                "\n" +
+                "class NotAvailable {\n" +
+                "    void testMethod(){\n" +
+                "        MockObject mockObject = Mockito.mock(MockObject.class);\n" +
+                "        BDDMockito.th<caret>en(mockObject);\n" +
+                "    }\n" +
+                "    private static class MockObject {\n" +
+                "        public void doSomething() {\n" +
+                "        }\n" +
+                "    }\n" +
+                "}");
+    }
+
+    public void testNotAvailableWhenSubsequentMethodCallToBDDMockitoThenIsNotShould() {
+        checkIntentionIsNotAvailable("NotAvailable.java",
+            "import org.mockito.BDDMockito;\n" +
+                "\n" +
+                "class NotAvailable {\n" +
+                "    void testMethod(){\n" +
+                "        MockObject mockObject = Mockito.mock(MockObject.class);\n" +
+                "        BDDMockito.th<caret>en(mockObject).shouldHaveNoInteractions();\n" +
+                "    }\n" +
+                "    private static class MockObject {\n" +
+                "        public void doSomething() {\n" +
+                "        }\n" +
+                "    }\n" +
+                "}");
+    }
+
+    public void testNotAvailableWhenBDDMockitoThenShouldHasNoSubsequentMethodCall() {
+        checkIntentionIsNotAvailable("NotAvailable.java",
+            "import org.mockito.BDDMockito;\n" +
+                "\n" +
+                "class NotAvailable {\n" +
+                "    void testMethod(){\n" +
+                "        MockObject mockObject = Mockito.mock(MockObject.class);\n" +
+                "        BDDMockito.th<caret>en(mockObject).should();\n" +
+                "    }\n" +
+                "    private static class MockObject {\n" +
+                "        public void doSomething() {\n" +
+                "        }\n" +
+                "    }\n" +
+                "}");
+    }
+
+    public void testAvailableForBDDMockitoThenShouldWithoutVerificationMode() {
+        checkIntentionIsAvailable("Available.java",
+            "import org.mockito.BDDMockito;\n" +
+                "\n" +
+                "class Available {\n" +
+                "    void testMethod(){\n" +
+                "        MockObject mockObject = Mockito.mock(MockObject.class);\n" +
+                "        BDDMockito.th<caret>en(mockObject).should().doSomething();\n" +
+                "    }\n" +
+                "    private static class MockObject {\n" +
+                "        public void doSomething() {\n" +
+                "        }\n" +
+                "    }\n" +
+                "}");
+    }
+
+    public void testAvailableForBDDMockitoThenShouldWithVerificationMode() {
+        checkIntentionIsAvailable("Available.java",
+            "import org.mockito.BDDMockito;\n" +
+                "import org.mockito.Mockito;\n" +
+                "\n" +
+                "class Available {\n" +
+                "    void testMethod(){\n" +
+                "        MockObject mockObject = Mockito.mock(MockObject.class);\n" +
+                "        BDDMockito.th<caret>en(mockObject).should(Mockito.times(2)).doSomething();\n" +
+                "    }\n" +
+                "    private static class MockObject {\n" +
+                "        public void doSomething() {\n" +
+                "        }\n" +
+                "    }\n" +
+                "}");
+    }
+
+    //Conversion
+
+    public void testConvertsBDDMockitoThenToMockitoVerifyWithoutVerificationMode() {
+        checkIntentionRun("ConversionTest.java",
+            "import org.mockito.BDDMockito;\n" +
+                "\n" +
+                "class ConversionTest {\n" +
+                "    void testMethod(){\n" +
+                "        MockObject mockObject = Mockito.mock(MockObject.class);\n" +
+                "        BDDMockito.th<caret>en(mockObject).should().doSomething();\n" +
+                "    }\n" +
+                "    private static class MockObject {\n" +
+                "        public void doSomething() {\n" +
+                "        }\n" +
+                "    }\n" +
+                "}",
+            "import org.mockito.BDDMockito;\n" +
+                "import org.mockito.Mockito;\n" +
+                "\n" +
+                "class ConversionTest {\n" +
+                "    void testMethod(){\n" +
+                "        MockObject mockObject = Mockito.mock(MockObject.class);\n" +
+                "        Mockito.verify(mockObject).doSomething();\n" +
+                "    }\n" +
+                "    private static class MockObject {\n" +
+                "        public void doSomething() {\n" +
+                "        }\n" +
+                "    }\n" +
+                "}");
+    }
+
+    public void testConvertsBDDMockitoThenToMockitoVerifyWithVerificationMode() {
+        checkIntentionRun("ConversionTest.java",
+            "import org.mockito.BDDMockito;\n" +
+                "import org.mockito.Mockito;\n" +
+                "\n" +
+                "class ConversionTest {\n" +
+                "    void testMethod(){\n" +
+                "        MockObject mockObject = Mockito.mock(MockObject.class);\n" +
+                "        BDDMockito.th<caret>en(mockObject).should(Mockito.times(2)).doSomething();\n" +
+                "    }\n" +
+                "    private static class MockObject {\n" +
+                "        public void doSomething() {\n" +
+                "        }\n" +
+                "    }\n" +
+                "}",
+            "import org.mockito.BDDMockito;\n" +
+                "import org.mockito.Mockito;\n" +
+                "\n" +
+                "class ConversionTest {\n" +
+                "    void testMethod(){\n" +
+                "        MockObject mockObject = Mockito.mock(MockObject.class);\n" +
+                "        Mockito.verify(mockObject, Mockito.times(2)).doSomething();\n" +
+                "    }\n" +
+                "    private static class MockObject {\n" +
+                "        public void doSomething() {\n" +
+                "        }\n" +
+                "    }\n" +
+                "}");
+    }
+
+    public void testConvertsBDDMockitoThenToMockitoVerifyWithVerificationModeWithLineWrapping() {
+        checkIntentionRun("ConversionTest.java",
+            "import org.mockito.BDDMockito;\n" +
+                "import org.mockito.Mockito;\n" +
+                "\n" +
+                "class ConversionTest {\n" +
+                "    void testMethod(){\n" +
+                "        MockObject mockObject = Mockito.mock(MockObject.class);\n" +
+                "        BDDMockito.th<caret>en(mockObject)\n" +
+                "            .should(Mockito.times(2)\n" +
+                "                .description(\"some description\"))\n" +
+                "            .doSomething();\n" +
+                "    }\n" +
+                "    private static class MockObject {\n" +
+                "        public void doSomething() {\n" +
+                "        }\n" +
+                "    }\n" +
+                "}",
+            "import org.mockito.BDDMockito;\n" +
+                "import org.mockito.Mockito;\n" +
+                "\n" +
+                "class ConversionTest {\n" +
+                "    void testMethod(){\n" +
+                "        MockObject mockObject = Mockito.mock(MockObject.class);\n" +
+                "        Mockito.verify(mockObject, Mockito.times(2)\n" +
+                "                .description(\"some description\"))\n" +
+                "            .doSomething();\n" +
+                "    }\n" +
+                "    private static class MockObject {\n" +
+                "        public void doSomething() {\n" +
+                "        }\n" +
+                "    }\n" +
+                "}");
+    }
+}